Got rid of extra directories in screens folder leaving only screen components; Creatad stand alone file for NavigationContainer consumed by App; All navigators live in their own files and are to be imported & nested as required; Root most navigator will be used as the root for NavigationContainer
