+++ app/routes/RootStack.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { createStackNavigator } from '@react-navigation/stack';
+
+const RootStack = createStackNavigator();
+
+export default function RootStack() {
+  return (
+      <RootStack.Navigator>
+        
+      </RootStack.Navigator>
+  );
+}
