+++ app/screens/MyExercises.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
-import CreateButton from '../../components/CreateButton'
-import Spacer from '../../components/Spacer';
+import CreateButton from '../components/CreateButton'
+import Spacer from '../components/Spacer';
 
 function MyExercises() {
   return (
