+++ app/screens/index.js
@@ -1,3 +1,3 @@
-export { default as Explore } from './Explore/ExploreTab'
-export { default as Library } from './Library/LibraryTab'
-export { default as Log } from './Log/LogTab'
+export { default as Explore } from './ExploreTab'
+export { default as Library } from './LibraryTab'
+export { default as Log } from './LogTab'
